fix: purchases/[id] folder
