+++ carbonmark-api/src/routes/purchases/[id]/get.ts
@@ -1,11 +1,9 @@
-import { utils } from "ethers";
 import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { isNil } from "lodash";
-import { Purchase } from "../../../models/Purchase.model";
-import { CreditId } from "../../../utils/CreditId";
+import { Purchase } from "src/models/Purchase.model";
 import { gql_sdk } from "../../../utils/gqlSdk";
+import { composePurchaseModel, isValidPurchaseId } from "../get.utils";
 import { Params, Querystring, schema } from "./get.schema";
-import { isValidPurchaseId } from "./get.utils";
 
 const handler = async (
   request: FastifyRequest<{
@@ -25,36 +23,17 @@ const handler = async (
     return reply.status(404).send({ error: "Purchase not found" });
   }
 
-  const project = purchase.listing.project;
-
-  const [, registryProjectId] = CreditId.splitProjectId(project.key);
-
-  const response: Purchase = {
-    id: purchase.id,
-    amount: utils.formatUnits(purchase.amount, 18),
-    price: utils.formatUnits(purchase.price, 6),
-    listing: {
-      id: purchase.listing.id,
-      tokenAddress: purchase.listing.tokenAddress,
-      seller: {
-        id: purchase.listing.seller.id,
-      },
-      project: {
-        key: purchase.listing.project.key,
-        vintage: purchase.listing.project.vintage,
-        methodology: purchase.listing.project.methodology,
-        name: project.name,
-        projectID: registryProjectId,
-        country: project.country.id,
-      },
-    },
-  };
+  const response = composePurchaseModel(purchase);
 
   return reply.status(200).send(response);
 };
 
 export default (fastify: FastifyInstance) =>
-  fastify.route({
+  fastify.route<{
+    Params: Params;
+    Querystring: Querystring;
+    Reply: Purchase | { error: string };
+  }>({
     method: "GET",
     url: "/purchases/:id",
     handler,
